requirements.txt
# Core dependencies
google-generativeai>=0.8.0
opencv-python>=4.8.0
mediapipe>=0.10.0
pytesseract>=0.3.10
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0

# Face recognition - using deepface instead of dlib-based face-recognition
# (dlib has compilation issues on macOS with Apple Silicon)
deepface>=0.0.89

# Audio processing (optional - remove if you don't need audio transcription)
# openai-whisper>=20231117

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional: API server
# fastapi>=0.104.0
# uvicorn>=0.24.0
